chore(meshwide): split on different files
