+++ plugins/lime-plugin-mesh-wide/src/components/Map.tsx
@@ -1,0 +1,21 @@
+import { MapContainer, TileLayer } from "react-leaflet";
+
+const openStreetMapTileString = "http://{s}.tile.osm.org/{z}/{x}/{y}.png";
+const openStreetMapAttribution =
+    '&copy; <a href="http://osm.org/copyright">OpenStreetMap</a> contributors';
+
+export const MeshWideMap = () => {
+    return (
+        <MapContainer
+            center={[-30, -60]}
+            zoom={3}
+            scrollWheelZoom={true}
+            className={"w-screen h-screen sm:h-auto sm:pt-14 z-0"}
+        >
+            <TileLayer
+                attribution={openStreetMapAttribution}
+                url={openStreetMapTileString}
+            />
+        </MapContainer>
+    );
+};
